Make grpc-replay runnable in tests
