+++ grpc-replay/grpc-replay.go
@@ -1,23 +1,11 @@
 package main
 
 import (
-	"encoding/json"
 	"flag"
 	"fmt"
-	"github.com/bradleyjkemp/grpc-tools/internal"
-	"github.com/bradleyjkemp/grpc-tools/internal/codec"
-	"github.com/bradleyjkemp/grpc-tools/internal/marker"
-	"github.com/bradleyjkemp/grpc-tools/internal/proto_decoder"
+	"github.com/bradleyjkemp/grpc-tools/grpc-replay/replay"
 	_ "github.com/bradleyjkemp/grpc-tools/internal/versionflag"
-	"github.com/sirupsen/logrus"
-	"golang.org/x/net/context"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials"
-	"google.golang.org/grpc/metadata"
-	"io"
 	"os"
-	"strings"
-	"time"
 )
 
 var (
@@ -29,126 +17,10 @@ var (
 
 func main() {
 	flag.Parse()
-	err := run()
+	err := replay.Run(*protoRoots, *protoDescriptors, *dumpPath, *destinationOverride)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err.Error())
 		flag.Usage()
 		os.Exit(1)
 	}
-}
-
-func run() error {
-	dumpFile, err := os.Open(*dumpPath)
-	if err != nil {
-		return err
-	}
-	var resolvers []proto_decoder.MessageResolver
-	if *protoRoots != "" {
-		r, err := proto_decoder.NewFileResolver(strings.Split(*protoRoots, ",")...)
-		if err != nil {
-			return err
-		}
-		resolvers = append(resolvers, r)
-	}
-	if *protoDescriptors != "" {
-		r, err := proto_decoder.NewDescriptorResolver(strings.Split(*protoRoots, ",")...)
-		if err != nil {
-			return err
-		}
-		resolvers = append(resolvers, r)
-	}
-	encoder := proto_decoder.NewEncoder(resolvers...)
-
-	dumpDecoder := json.NewDecoder(dumpFile)
-RPC:
-	for {
-		rpc := internal.RPC{}
-		err := dumpDecoder.Decode(&rpc)
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			return fmt.Errorf("failed to decode dump: %s", err)
-		}
-
-		conn, err := getConnection(rpc.Metadata)
-		if err != nil {
-			return fmt.Errorf("failed to connect to destination (%s): %s", *destinationOverride, err)
-		}
-
-		// RPC has metadata added by grpc-dump that should be removed before sending
-		// (so that we're sending as close as possible to the original request)
-		marker.RemoveHTTPSMarker(rpc.Metadata)
-
-		ctx := metadata.NewOutgoingContext(context.Background(), rpc.Metadata)
-		streamName := rpc.StreamName()
-		str, err := conn.NewStream(ctx, &grpc.StreamDesc{
-			StreamName:    streamName,
-			ServerStreams: true,
-			ClientStreams: true,
-		}, streamName)
-		if err != nil {
-			return fmt.Errorf("failed to make new stream: %v", err)
-		}
-
-		fmt.Print(streamName, "...")
-		for _, message := range rpc.Messages {
-			msgBytes, err := encoder.Encode(streamName, message)
-			if err != nil {
-				return fmt.Errorf("failed to encode message: %v", err)
-			}
-
-			switch message.MessageOrigin {
-			case internal.ClientMessage:
-				err := str.SendMsg(msgBytes)
-				if err != nil {
-					return fmt.Errorf("failed to send message: %v", err)
-				}
-			case internal.ServerMessage:
-				var resp []byte
-				err := str.RecvMsg(&resp)
-				if err != nil {
-					// TODO when do we assert on RPC errors?
-					return fmt.Errorf("failed to recv message: %v", err)
-				}
-				if string(resp) != string(msgBytes) {
-					fmt.Println("Err mismatch")
-					continue RPC
-				}
-			default:
-				return fmt.Errorf("invalid message type: %v", message.MessageOrigin)
-			}
-		}
-		fmt.Println("OK")
-	}
-	return nil
-}
-
-var cachedConns = internal.NewConnPool(logrus.New())
-
-func getConnection(md metadata.MD) (*grpc.ClientConn, error) {
-	// if no destination override set then auto-detect from the metadata
-	var destination = *destinationOverride
-	if destination == "" {
-		authority := md.Get(":authority")
-		if len(authority) == 0 {
-			return nil, fmt.Errorf("no destination override specified and could not auto-detect from dump")
-		}
-		destination = authority[0]
-	}
-
-	options := []grpc.DialOption{
-		grpc.WithDefaultCallOptions(grpc.ForceCodec(codec.NoopCodec{})),
-		grpc.WithBlock(),
-	}
-
-	if marker.IsTLSRPC(md) {
-		options = append(options, grpc.WithTransportCredentials(credentials.NewTLS(nil)))
-	} else {
-		options = append(options, grpc.WithInsecure())
-	}
-
-	dialCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	return cachedConns.GetClientConn(dialCtx, destination, options...)
 }
